requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
ccxt>=4.0.0
python-binance>=1.0.0

# ML and Data Science
scikit-learn>=1.3.0
tensorflow>=2.13.0
xgboost>=1.7.0
lightgbm>=4.0.0

# Data processing and analysis
ta>=0.10.0  # Technical Analysis
yfinance>=0.2.0
requests>=2.31.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
sqlite3
backtrader>=1.9.0

# Configuration and logging
python-dotenv>=1.0.0
pyyaml>=6.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Web framework (for future dashboard)
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.25.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Async support
asyncio
aiohttp>=3.8.0
websockets>=11.0.0

# Date and time
python-dateutil>=2.8.0
pytz>=2023.3